+++ R/IntCal.R
@@ -14,11 +14,11 @@ NULL
 
 # todo: add another caldist function a la l.calib of coffee for single cal yrs but multiple dates? hpds sometimes are >100%??? prepare calib function with MCMC ccurve.
 
-# done: repaired problems with depths and calheights in draw.dates, draw.ccurve now plots correct label when BCAD, c14.lim now estimated correctly when a second curve is added
+# done: 
 
 
 #' @name list.ccurves
-#' @title List the calibration curves.
+#' @title List the calibration curves
 #' @description List the file names of the calibration curves available within the IntCal package.
 #' @export
 list.ccurves <- function() {
@@ -66,6 +66,8 @@ copyCalibrationCurve <- function(cc=1, postbomb=FALSE) {
 #' Hogg et al. 2020 SHCal20 Southern Hemisphere calibration, 0-55,000 years cal BP. Radiocarbon 62. \doi{10.1017/RDC.2020.59}
 #'
 #' Hua et al. 2013 Atmospheric radiocarbon for the period 1950-2010. Radiocarbon 55(4), \doi{10.2458/azu_js_rc.v55i2.16177}
+#' 
+#' Hua et al. 2021 Atmospheric radiocarbon for the period 1950-2019. Radiocarbon in press,  \doi{10.1017/RDC.2021.95}
 #'
 #' Hughen et al. 2020 Marine20-the marine radiocarbon age calibration curve (0-55,000 cal BP). Radiocarbon 62. \doi{10.1017/RDC.2020.68}
 #'