+++ NEWS.md
@@ -2,6 +2,7 @@
 * draw.ccurve now plots the correct label when using BCAD
 * draw.ccurve now plots depths at the expected heights
 * draw.dates now plots multiple dates at the expected heights (with more precise dates peaking higher)
+* corrected the Rmd files which had erroneous formatting and some confusing examples
 
 # IntCal 0.2.2
 * in mix.curves(), calibration curves are now written to a temporary directory by default, as per CRAN policies