+++ NEWS.md
@@ -1,6 +1,5 @@
 # IntCal 0.2.3
-* draw.ccurve now plots the correct label when using BCAD
-* draw.ccurve now plots depths at the expected heights
+* draw.ccurve now plots the correct label when using BCAD, and plots depths at the expected heights
 * draw.dates now plots multiple dates at the expected heights (with more precise dates peaking higher)
 * corrected the Rmd files which had erroneous formatting and some confusing examples